Add a `withTransition` decorator
