+++ packages/ui/decorators/index.js
@@ -1,0 +1,1 @@
+export { default as withTransition } from './withTransition.js';
