+++ packages/ui/index.js
@@ -2,3 +2,4 @@ export * from './primitives/index.js';
 export * from './atoms/index.js';
 export * from './molecules/index.js';
 export * from './organisms/index.js';
+export * from './decorators/index.js';
